requirements.txt
# AutistBoar Python dependencies
# Install: pip install -r requirements.txt

# HTTP client
httpx>=0.27

# Config & data models
pydantic>=2.6
pyyaml>=6.0
python-dotenv>=1.0

# Solana
solders>=0.21
solana>=0.34

# Edge Bank (vector recall)
sentence-transformers>=3.0
sqlite-vec>=0.1

# Testing
pytest>=8.0
pytest-asyncio>=0.23
